fix: store refactor
