+++ client/apps/eternum-mobile/src/main.tsx
@@ -10,8 +10,8 @@ import { dojoConfig } from "../dojoConfig";
 import { env } from "../env";
 import { initialSync } from "./app/dojo/sync";
 import { ETERNUM_CONFIG } from "./shared/config/config";
-import { useStore } from "./shared/store/use-store";
-import { WorldStore } from "./shared/store/use-world-loading";
+import { useStore } from "./shared/store";
+import { WorldSlice } from "./shared/store/slices/world-loading-slice";
 
 async function main() {
   const setupResult = await setup(
@@ -33,7 +33,7 @@ async function main() {
 
   const eternumConfig = ETERNUM_CONFIG();
   configManager.setDojo(setupResult.components, eternumConfig);
-  const state: WorldStore = useStore.getState();
+  const state: WorldSlice = useStore.getState();
   await initialSync(setupResult, state);
 
   createRoot(document.getElementById("root")!).render(
