+++ client/apps/eternum-mobile/src/shared/consts/index.ts
@@ -1,0 +1,2 @@
+export * from "./constants";
+export * from "./routes";
